Linux specific fix for switching languages: PrusaSlicer 2.3 forces en_GB locale when switching to English from the user interface, while often Linux users have just en_US locales configured.
With this commit the user will have to select English (U.S.) or
English (U.K.). Usually English (U.S.) will work.

Vojtech is scared to touch the code for the other platforms (Windows and Mac)
as the language switching has been reworked multiple times and it is quite
fragile.