creality.ini: enable conservative elefant_foot_compensation
given the diversity of printers and bed types and a typical lack of
auto bed levelling, it will be impossible to get a good fit
for everyone by default.

A very conservative 0.1 for elefant_foot_compensation is unlikely
to cause overcompensation.

And I'd prefer to undercompensate as opposed to overcompensate
as a default.