Windows specific: Use mipmaps generated through OpenGL for AMD graphics cards with drivers newer than 22.6.1.
Since AMD driver version 22.7.1, there is probably some bug in the driver that causes the issue with the missing texture of the bed. It seems that this issue only triggers when mipmaps are generated manually (combined with a texture compression) and when mipmaps are generated through OpenGL glGenerateMipmap is working.

This workaround detects the AMD driver version and generates mipmaps through OpenGL for driver versions newer than 22.6.1.